LayoutTests/platform/mac-mojave/fast/text/international/system-language/navigator-language/navigator-language-ja-expected.txt

FAIL navigator.language should be ja. Was ja-jp.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
